+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## [1.1.6](https://github.com/jumpei999/portfolio/compare/v1.1.5...v1.1.6) (2026-07-04)
+
+
+### Bug Fixes
+
+* fix history timeline click bounce ([2209cba](https://github.com/jumpei999/portfolio/commit/2209cbaca7ff83730c466986dbc704dfb30c52df))
+* fix history timeline synchronization on mobile ([d92b9e2](https://github.com/jumpei999/portfolio/commit/d92b9e207b196ff150cc3cc099e12b4fbfe4d534))
+
 ## [1.1.5](https://github.com/jumpei999/portfolio/compare/v1.1.4...v1.1.5) (2026-06-30)